[Move Only] Move wallet related things to src/wallet/
