+++ src/init.cpp
@@ -30,9 +30,9 @@
 #include "smessage.h"
 
 #ifdef ENABLE_WALLET
-#include "db.h"
-#include "wallet.h"
-#include "walletdb.h"
+#include "wallet/db.h"
+#include "wallet/wallet.h"
+#include "wallet/walletdb.h"
 #endif
 
 #include <boost/algorithm/string/predicate.hpp>
