+++ src/kernelrecord.cpp
@@ -1,6 +1,6 @@
 #include "kernelrecord.h"
 
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "base58.h"
 
 using namespace std;
