+++ src/net.cpp
@@ -11,7 +11,7 @@
 #include "primitives/transaction.h"
 #include "ui_interface.h"
 #include "darksend.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #ifdef WIN32
 #include <string.h>
