+++ src/qt/addresstablemodel.cpp
@@ -3,7 +3,7 @@
 #include "guiutil.h"
 #include "walletmodel.h"
 
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "base58.h"
 #include "stealth.h"
