+++ src/qt/blockbrowser.cpp
@@ -1,7 +1,7 @@
 #include "blockbrowser.h"
 #include "ui_blockbrowser.h"
 #include "main.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "base58.h"
 #include "clientmodel.h"
 #include "walletmodel.h"
