+++ src/qt/optionsmodel.cpp
@@ -11,8 +11,8 @@
 #include "main.h"
 #include "net.h"
 #ifdef ENABLE_WALLET
-#include "wallet.h"
-#include "walletdb.h"
+#include "wallet/wallet.h"
+#include "wallet/walletdb.h"
 #endif
 #include "netbase.h"
