+++ src/qt/sendcoinsdialog.cpp
@@ -15,7 +15,7 @@
 #include "walletmodel.h"
 #include "base58.h"
 #include "coincontrol.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "init.h"
 #include "addressbookpage.h"
 #include "askpassphrasedialog.h"
