+++ src/qt/signverifymessagedialog.cpp
@@ -8,7 +8,7 @@
 #include "main.h"
 #include "optionsmodel.h"
 #include "walletmodel.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <QClipboard>
