+++ src/qt/transactiondesc.cpp
@@ -9,7 +9,7 @@
 #include "transactionrecord.h"
 #include "util.h"
 #include "ui_interface.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "txdb.h"
 
 #include <string>
