+++ src/qt/transactionrecord.cpp
@@ -2,7 +2,7 @@
 
 #include "base58.h"
 #include "util.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "darksend.h"
 #include "instantx.h"
