+++ src/qt/transactiontablemodel.cpp
@@ -9,7 +9,7 @@
 #include "addresstablemodel.h"
 #include "bitcoinunits.h"
 
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "ui_interface.h"
 
 #include <QList>
