+++ src/qt/walletmodel.cpp
@@ -17,8 +17,8 @@
 #include "keystore.h"
 #include "main.h"
 #include "ui_interface.h"
-#include "wallet.h"
-#include "walletdb.h" // for BackupWallet
+#include "wallet/wallet.h"
+#include "wallet/walletdb.h" // for BackupWallet
 #include "spork.h"
 #include "smessage.h"
