+++ src/rpcmisc.cpp
@@ -13,8 +13,8 @@
 #include "stealth.h"
 #include "spork.h"
 #ifdef ENABLE_WALLET
-#include "wallet.h"
-#include "walletdb.h"
+#include "wallet/wallet.h"
+#include "wallet/walletdb.h"
 #endif
 
 #include <stdint.h>
