+++ src/smessage.h
@@ -10,7 +10,7 @@
 
 #include "net.h"
 #include "db.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "base58.h"
 #include "lz4/lz4.h"
